+++ data/src/main/java/com/petid/data/api/HospitalAPI.kt
@@ -42,7 +42,7 @@ interface HospitalAPI {
      */
     @GET("/v1/hospital/order/time")
     suspend fun getHospitalOrderTimeList(
-        @Query("hospitalId") hospitalId: Int,
+        @Query("hospitalId") hospitalId: Long,
         @Query("day") day: String,
         @Query("date") date: String // yyyy-mm-dd
     ) : List<String>
@@ -77,7 +77,7 @@ interface HospitalAPI {
      */
     @DELETE("/v1/hospital/order/{orderId}")
     suspend fun deleteHospitalOrder(
-        @Path("orderId") orderId: Int,
+        @Path("orderId") orderId: Long,
     ) : Int
 
     /**
@@ -93,6 +93,6 @@ interface HospitalAPI {
      */
     @GET("/v1/hospital/{hospitalId}")
     suspend fun getHospitalDetailById(
-        @Path("hospitalId") hospitalId: Int,
+        @Path("hospitalId") hospitalId: Long,
     ) : HospitalResponse
 }
