+++ domain/src/main/java/com/petid/domain/repository/ReservationHistoryInfoRepository.kt
@@ -5,5 +5,5 @@ import com.petid.domain.util.ApiResult
 
 interface ReservationHistoryInfoRepository {
     suspend fun getHospitalReservationHistoryList(status: String): ApiResult<List<HospitalOrderDetailEntity>>
-    suspend fun cancelHospitalReservation(orderId: Int): ApiResult<Int>
+    suspend fun cancelHospitalReservation(orderId: Long): ApiResult<Int>
 }
